Day 15: using int64 instead of big.Int...
I actualy forgot the name of the int64 when I was coding this.

+++ day15/part1/generators/main_test.go
@@ -8,9 +8,9 @@ import (
 
 var generatorNextTT = []struct {
 	name             string
-	memA, memB       int
-	factorA, factorB int
-	expected         int
+	memA, memB       int64
+	factorA, factorB int64
+	expected         int64
 }{
 	{"provided example", 65, 8921, 16807, 48271, 588},
 }
